+++ espnet2/gan_tts/hifigan/__init__.py
@@ -1,0 +1,6 @@
+from espnet2.gan_tts.hifigan.hifigan import HiFiGANGenerator  # NOQA
+from espnet2.gan_tts.hifigan.hifigan import HiFiGANMultiPeriodDiscriminator  # NOQA
+from espnet2.gan_tts.hifigan.hifigan import HiFiGANMultiScaleDiscriminator  # NOQA
+from espnet2.gan_tts.hifigan.hifigan import HiFiGANMultiScaleMultiPeriodDiscriminator  # NOQA
+from espnet2.gan_tts.hifigan.hifigan import HiFiGANPeriodDiscriminator  # NOQA
+from espnet2.gan_tts.hifigan.hifigan import HiFiGANScaleDiscriminator  # NOQA
