+++ espnet2/gan_tts/hifigan/hifigan.py
@@ -18,7 +18,7 @@ from typing import Optional
 import torch
 import torch.nn.functional as F
 
-from espnet2.gan_tts.vits.residual_block import HiFiGANResidualBlock as ResidualBlock
+from espnet2.gan_tts.hifigan.residual_block import ResidualBlock
 
 
 class HiFiGANGenerator(torch.nn.Module):
