+++ espnet2/gan_tts/vits/generator.py
@@ -19,7 +19,7 @@ import torch.nn.functional as F
 
 from espnet.nets.pytorch_backend.nets_utils import make_non_pad_mask
 from espnet2.gan_tts.vits.duration_predictor import StochasticDurationPredictor
-from espnet2.gan_tts.vits.hifigan import HiFiGANGenerator
+from espnet2.gan_tts.hifigan import HiFiGANGenerator
 from espnet2.gan_tts.vits.posterior_encoder import PosteriorEncoder
 from espnet2.gan_tts.vits.residual_coupling import ResidualAffineCouplingBlock
 from espnet2.gan_tts.vits.text_encoder import TextEncoder
