+++ espnet2/gan_tts/vits/posterior_encoder.py
@@ -13,8 +13,8 @@ from typing import Tuple
 import torch
 
 from espnet.nets.pytorch_backend.nets_utils import make_non_pad_mask
-from espnet2.gan_tts.vits.residual_block import Conv1d
-from espnet2.gan_tts.vits.wavenet import WaveNet
+from espnet2.gan_tts.wavenet.residual_block import Conv1d
+from espnet2.gan_tts.wavenet import WaveNet
 
 
 class PosteriorEncoder(torch.nn.Module):
