+++ espnet2/gan_tts/wavenet/__init__.py
@@ -1,0 +1,1 @@
+from espnet2.gan_tts.wavenet.wavenet import WaveNet  # NOQA
